pyproject.toml
[project]
name = "tts_generator"
version = "0.1.0"
description = "Simple text-to-speech generator using multiple providers"
requires-python = ">=3.8"
dependencies = [
    "gTTS>=2.3.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "typer>=0.9.0",
    "python-slugify",
]

[project.scripts]
tts = "src.tts_generator:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]  # Changed to include the whole src directory
